requirements.txt
Appium-Python-Client~=4.0.1
selenium~=4.23.1
pytest~=8.3.2
pytest-html
requests==2.32.3
testrail-api==1.13.2
py==1.11.0
pytest-metadata==1.11.0
pytest-xdist

PyYAML~=6.0.1